showcase_local_coffee_shop/google_cloud_run/trigger_cloud_run_job/requirements.txt
functions-framework==3.4.0
google-auth
requests